redux users state implemented
