+++ webpack_src/src/app/components/Header/index.jsx
@@ -8,13 +8,13 @@ import FlatButton from 'material-ui/FlatButton';
 import AuthForm from 'app/components/AuthForm';
 
 import style from './style.scss'
-// { isAuthenticated,  handleLogout, ...props }
+
 const Header = (props) => {
 	return (
 		<AppBar
-			title={props.user.username ? `RestApp: Welcome, ${props.user.username}` : 'RestApp'}
+			title={props.user ? `RestApp: Welcome, ${props.users[props.user].username}` : 'RestApp'}
 			iconElementRight={
-				props.user.username ?
+				props.user ?
 					<FlatButton
 						label="LogOut"
 						onTouchTap={props.handleLogout}
@@ -29,7 +29,8 @@ const Header = (props) => {
 
 const mapStateToProps = (state) => {
 	return {
-		user: state.user
+		user: state.common.user,
+		users: state.users
 	}
 }
