+++ webpack_src/src/app/reducers/index.js
@@ -3,13 +3,13 @@ import commonSet from './commonSet';
 import mainpage from './mainpage';
 import common from './common';
 import dialogs from './dialogs';
-import user from './user';
+import users from './users';
 import { reducer as formReducer } from 'redux-form'
 
 export default combineReducers({
 	mainpage,
 	common,
 	dialogs,
-	user,
+	users,
 	form: formReducer
 })
